feat: PRE_DEPENDENCIES_SOLVING listener
Attempt at updating the package to rewrite prior to writing the
`composer.json`.

This _does_ work to a degree: it changes the job to reflect a
transformed package name, and, in fact means that the
onPrePackageInstall command is able to exit early. HOWEVER:

- I'm not checking to see if an install or update operation was
  requested. As such, this will also affect things like remove
  operations, or "why" operations.
- The `composer.json` still reflects the _requested_ package, not the
  _replacement_ package.

It looks like the `pre-command-run` option may be what we want here,
though this event listener should likely still be used for purposes of
`install` or `update` operations when the `composer.lock` file does not
yet exist.