fix: ensure replacement package is installed
In testing, we discovered that while composer reports installing Laminas
packages for nested dependencies, the actual packages installed were ZF
packages. (Determined by examining the contents of the vendor
directory.)

The solution was to update the `OperationInterface` with the replacement
package found via the repository manager, instead of updating the
package with the desired name. Replacing the package instance in the
operation ensures that the actual package is installed.

One note: packages slip-streamed in this way do not return meaningful
information via `composer why`. I am not certain there is a way to make
that happen.